Move helpActivity to detected package
